Wall Collision & Switch Images Function
Wall collision is working properly now. THere are four rays that
go in the forward, backwards, left and right direction. The
rotation matrix was needed for the left and righ directions.
Added a new function that switches from the drawn image to
the image that is going to be displaid.