Mouse Rotation
Incorportated a function that detects the movement of the pointer
on the screen. It checks whether the difference between the curr-
ent and the previous position of the mouse is positive (right) or
negative (left). So you do not move like a crazy bastard, the
absoulte value of delta_x is compared to the threshold of the
mouse sensitivity. If that is the case, then you wont move.